hwdb: Add Kensington Expert Mouse (optical; model # K64325)
Resolution measured by marking the ball and rolling it over a whole
number of times, then dividing by a suitable multiple of the ball's
circumference.  This gave results between 418 and 423 DPI, so go for
420.

Scroll angle measured much the same way, but the trackball seems to lose
counts if the scroll ring is moved too quickly.  I think I got a maximum
of 40 per revolution, but I'm not entirely sure.